Moved matrix gui code to UI package
* Minor refactor of delete items to not repeat code
