+++ src/main/java/Matrices/EditorTabs/MultiDomainEditorTab.java
@@ -1,18 +1,24 @@
 package Matrices.EditorTabs;
 
-import Matrices.Data.*;
+import Matrices.Data.AbstractDSMData;
+import Matrices.Data.AsymmetricDSMData;
 import Matrices.Data.Entities.Grouping;
+import Matrices.Data.MultiDomainDSMData;
+import Matrices.Data.SymmetricDSMData;
 import Matrices.IOHandlers.AbstractIOHandler;
 import Matrices.IOHandlers.AsymmetricIOHandler;
 import Matrices.IOHandlers.MultiDomainIOHandler;
 import Matrices.IOHandlers.SymmetricIOHandler;
-import Matrices.SideBarTools.AbstractSideBar;
-import Matrices.SideBarTools.AsymmetricSideBar;
-import Matrices.SideBarTools.MultiDomainSideBar;
-import Matrices.SideBarTools.SymmetricSideBar;
-import Matrices.Views.*;
 import UI.HeaderMenu;
 import UI.MatrixMetaDataPane;
+import UI.MatrixViews.AbstractMatrixView;
+import UI.MatrixViews.AsymmetricView;
+import UI.MatrixViews.MultiDomainView;
+import UI.MatrixViews.SymmetricView;
+import UI.SideBarViews.AbstractSideBar;
+import UI.SideBarViews.AsymmetricSideBar;
+import UI.SideBarViews.MultiDomainSideBar;
+import UI.SideBarViews.SymmetricSideBar;
 import UI.Widgets.DraggableTab;
 import UI.Widgets.Misc;
 import javafx.geometry.Insets;
