+++ src/main/java/UI/ClusterAlgorithmWindow.java
@@ -4,13 +4,12 @@ import Matrices.ClusterAlgorithms.Thebeau;
 import Matrices.Data.Entities.Grouping;
 import Matrices.Data.SymmetricDSMData;
 import Matrices.IOHandlers.SymmetricIOHandler;
-import Matrices.Views.AbstractMatrixView;
-import Matrices.Views.SymmetricView;
+import UI.MatrixViews.AbstractMatrixView;
+import UI.MatrixViews.SymmetricView;
 import UI.ClusterAlgorithmViews.ART1View;
 import UI.ClusterAlgorithmViews.IAlgorithmView;
 import UI.ClusterAlgorithmViews.ParameterBuilder;
 import UI.ClusterAlgorithmViews.ThebeauView;
-import UI.Widgets.NumericTextField;
 import javafx.application.Platform;
 import javafx.beans.property.*;
 import javafx.collections.FXCollections;
