+++ src/main/java/UI/MatrixViews/Flags/ISymmetricHighlight.java
@@ -1,4 +1,4 @@
-package Matrices.Views.Flags;
+package UI.MatrixViews.Flags;
 
 
 /**
