+++ src/main/java/UI/MatrixViews/SymmetricView.java
@@ -1,12 +1,13 @@
-package Matrices.Views;
+package UI.MatrixViews;
 
 import Constants.Constants;
-import Matrices.Data.AsymmetricDSMData;
 import Matrices.Data.Entities.DSMConnection;
 import Matrices.Data.Entities.DSMItem;
 import Matrices.Data.Entities.Grouping;
 import Matrices.Data.Entities.RenderMode;
-import Matrices.Views.Entities.Cell;
+import Matrices.Data.SymmetricDSMData;
+import UI.MatrixViews.Entities.Cell;
+import UI.MatrixViews.Flags.ISymmetricHighlight;
 import UI.Widgets.FreezeGrid;
 import UI.Widgets.Misc;
 import UI.Widgets.NumericTextField;
@@ -25,23 +26,17 @@ import javafx.scene.control.ListView;
 import javafx.scene.input.MouseButton;
 import javafx.scene.layout.*;
 import javafx.scene.paint.Color;
-import javafx.scene.paint.CycleMethod;
-import javafx.scene.paint.LinearGradient;
-import javafx.scene.paint.Stop;
-import javafx.scene.text.Font;
 import javafx.util.Callback;
 import javafx.util.Pair;
 import org.javatuples.Triplet;
 
 import java.util.*;
 
+public class SymmetricView extends AbstractMatrixView implements ISymmetricHighlight {
 
-/**
- * class for viewing and editing asymmetric matrices
- */
-public class AsymmetricView extends AbstractMatrixView {
+    private boolean symmetryValidation = false;
 
-    AsymmetricDSMData matrix;
+    SymmetricDSMData matrix;
 
     /**
      * Returns a MatrixGuiHandler object for a given matrix
@@ -49,7 +44,7 @@ public class AsymmetricView extends AbstractMatrixView {
      * @param matrix   the SymmetricDSMData object to display
      * @param fontSize the default font size to display the matrix with
      */
-    public AsymmetricView(AsymmetricDSMData matrix, double fontSize) {
+    public SymmetricView(SymmetricDSMData matrix, double fontSize) {
         super(matrix, fontSize);
         this.matrix = matrix;
     }
@@ -61,7 +56,7 @@ public class AsymmetricView extends AbstractMatrixView {
      * @param fontSize  the new font size for the matrix view
      * @return          this
      */
-    public AsymmetricView withFontSize(double fontSize) {
+    public SymmetricView withFontSize(double fontSize) {
         this.fontSize.set(fontSize);
         return this;
     }
@@ -73,21 +68,21 @@ public class AsymmetricView extends AbstractMatrixView {
      * @param mode  the new mode for the matrix view
      * @return      this
      */
-    public AsymmetricView withMode(MatrixViewMode mode) {
+    public SymmetricView withMode(MatrixViewMode mode) {
         this.currentMode.set(mode);
         return this;
     }
 
 
     /**
-     * Copy constructor for AsymmetricView class. Performs a deep copy on the matrix data. Everything else will be
+     * Copy constructor for SymmetricView class. Performs a deep copy on the matrix data. Everything else will be
      * generated when the refreshView method is called
      *
-     * @return  the copy of the current AsymmetricView
+     * @return  the copy of the current SymmetricView
      */
     @Override
-    public AsymmetricView createCopy() {
-        AsymmetricView copy = new AsymmetricView(matrix.createCopy(), fontSize.doubleValue());
+    public SymmetricView createCopy() {
+        SymmetricView copy = new SymmetricView(matrix.createCopy(), fontSize.doubleValue());
 
         copy.setCurrentMode(getCurrentMode());
 
@@ -98,6 +93,7 @@ public class AsymmetricView extends AbstractMatrixView {
     }
 
 
+    //region highlight methods
     /**
      * updates the background color of a cell based on the backgrounds set for it. Error highlight
      * is given the highest priority, then cross highlighting, then user highlighting, and lastly the grouping
@@ -126,7 +122,7 @@ public class AsymmetricView extends AbstractMatrixView {
             cell.setCellTextColor(Grouping.DEFAULT_FONT_COLOR);
 
         } else {  // default background determined by groupings
-            Integer rowUid = getUidsFromGridLoc(cell.getGridLocation()).getKey();
+            Integer rowUid = getUidsFromGridLoc(cell.getGridLocation()).getKey();  // null is used to check if it is an item or grouping cell
             Integer colUid = getUidsFromGridLoc(cell.getGridLocation()).getValue();
             if (rowUid == null && colUid != null) {  // highlight with column color
                 cell.setCellHighlight(matrix.getColItem(colUid).getGroup1().getColor());
@@ -134,23 +130,118 @@ public class AsymmetricView extends AbstractMatrixView {
             } else if (rowUid != null && colUid == null) {  // highlight with row color
                 cell.setCellHighlight(matrix.getRowItem(rowUid).getGroup1().getColor());
                 cell.setCellTextColor(matrix.getRowItem(rowUid).getGroup1().getFontColor());
-            } else if (rowUid != null && colUid != null) {  // highlight with row group color col group color
-                Stop[] stops = new Stop[] { new Stop(0, matrix.getRowItem(rowUid).getGroup1().getColor()), new Stop(1, matrix.getColItem(colUid).getGroup1().getColor())};
-                LinearGradient lg1 = new LinearGradient(0, 0, 1, 1, true, CycleMethod.NO_CYCLE, stops);
-
-                cell.setCellHighlight(new Background(new BackgroundFill(lg1, new CornerRadii(3), new Insets(0))));  // row and column color will be the same because row and column
+            } else if (
+                    rowUid != null && colUid != null
+                    && !rowUid.equals(matrix.getColItem(colUid).getAliasUid())
+                    && matrix.getRowItem(rowUid).getGroup1().equals(matrix.getColItem(colUid).getGroup1())
+            ) {
+                // row and column color will be the same because row and column
+                // have same group in symmetric matrix
+                cell.setCellHighlight(matrix.getRowItem(rowUid).getGroup1().getColor());
                 cell.setCellTextColor(matrix.getRowItem(rowUid).getGroup1().getFontColor());
             } else {
                 cell.setCellHighlight(cell.getHighlightBG("default"));
                 cell.setCellTextColor(Grouping.DEFAULT_FONT_COLOR);
+            }
+
+        }
+    }
+
+
+
+    /**
+     * Sets or clears a cells symmetry highlight based on the symmetryValidation flag
+     *
+     * @param cell  the cell to check the highlighting for
+     */
+    private void symmetryHighlightCell(Cell cell) {
+        Pair<Integer, Integer> gridLocation = cell.getGridLocation();
+        Pair<Integer, Integer> uids = getUidsFromGridLoc(gridLocation);
+        if(uids.getKey() == null || uids.getValue() == null) {
+            return;
+        }
+
+        int rowUid = uids.getKey();
+        int colUid = uids.getValue();
+        DSMConnection conn = matrix.getConnection(rowUid, colUid);
+        Pair<Integer, Integer> symmetricConnUids = matrix.getSymmetricConnectionUids(rowUid, colUid);
+        DSMConnection symmetricConn = matrix.getConnection(symmetricConnUids.getKey(), symmetricConnUids.getValue());
+
+        if(symmetryValidation && ((conn == null && symmetricConn != null) || (conn != null && symmetricConn == null) || (conn != null && symmetricConn != null && !conn.isSameConnectionType(symmetricConn)))) {
+            this.setCellHighlight(cell, AbstractMatrixView.SYMMETRY_ERROR_BACKGROUND, "symmetryError");
+            this.setCellHighlight(this.getGridLocFromUids(symmetricConnUids), AbstractMatrixView.SYMMETRY_ERROR_BACKGROUND, "symmetryError");
+        } else {
+            this.clearCellHighlight(cell, "symmetryError");
+            this.clearCellHighlight(this.getGridLocFromUids(symmetricConnUids), "symmetryError");
+        }
+    }
+
+
+    /**
+     * Highlights all cells symmetrically, faster than using symmetryHighlightCell in a loop because it will not double
+     * check the highlight of all cells
+     */
+    private void symmetryHighlightAllCells() {
+        for(Cell cell : cells) {
+            Pair<Integer, Integer> gridLocation = cell.getGridLocation();
+            Pair<Integer, Integer> uids = getUidsFromGridLoc(gridLocation);
+            if(uids.getKey() == null || uids.getValue() == null) {
+                continue;
+            }
+
+            int rowUid = uids.getKey();
+            int colUid = uids.getValue();
+            DSMConnection conn = matrix.getConnection(rowUid, colUid);
+            Pair<Integer, Integer> symmetricConnUids = matrix.getSymmetricConnectionUids(rowUid, colUid);
+            DSMConnection symmetricConn = matrix.getConnection(symmetricConnUids.getKey(), symmetricConnUids.getValue());
+
+            // ignore the symmetric connection because it will be hit later in the loop
+            if(symmetryValidation && ((conn == null && symmetricConn != null) || (conn != null && symmetricConn == null) || (conn != null && symmetricConn != null && !conn.isSameConnectionType(symmetricConn)))) {
+                this.setCellHighlight(cell, AbstractMatrixView.SYMMETRY_ERROR_BACKGROUND, "symmetryError");
+            } else {
+                this.clearCellHighlight(cell, "symmetryError");
             }
         }
     }
 
 
     /**
-     * Creates the gui that displays a matrix. Uses the AsymmetricDSMData's getGridArray() method to create the grid.
-     * Puts grid in a scroll pane and adds a location label (displays connection row, column) at the bottom of the VBox.
+     * Sets symmetryValidation to true in order to highlight symmetry errors
+     */
+    public void setValidateSymmetry() {
+        symmetryValidation = true;
+        symmetryHighlightAllCells();
+        for(Cell cell : cells) {
+            refreshCellHighlight(cell);
+        }
+    }
+
+
+    /**
+     * Sets symmetryValidation to false in order to stop highlighting symmetry errors
+     */
+    public void clearValidateSymmetry() {
+        symmetryValidation = false;
+        symmetryHighlightAllCells();
+        for(Cell cell : cells) {
+            refreshCellHighlight(cell);
+        }
+    }
+
+
+    /**
+     * @return  if validation symmetry is set or not
+     */
+    public boolean getSymmetryValidation() {
+        return symmetryValidation;
+    }
+    //endregion
+
+
+    /**
+     * Creates the gui that displays a matrix with an editable view. Uses the SymmetricDSMData's getGridArray() method
+     * to create the grid. Adds a location label (displays connection row, column)
+     * at the bottom of the VBox.
      */
     @Override
     protected void refreshEditView() {
@@ -175,7 +266,7 @@ public class AsymmetricView extends AbstractMatrixView {
         ComboBox<Grouping> _groupings = new ComboBox<>();
         _groupings.setMinWidth(Region.USE_PREF_SIZE);
         _groupings.setPadding(new Insets(0));
-        _groupings.setStyle("-fx-background-color: transparent; -fx-padding: 0, 0, 0, 0; -fx-font-size: " + (fontSize.doubleValue()) + " };");
+        _groupings.setStyle("-fx-background-color: transparent; -fx-padding: 0, 0, 0, 0; -fx-font-size: " + (fontSize.doubleValue()) + " ;");
         double maxHeight = Misc.calculateNodeSize(_groupings).getHeight();
 
 
@@ -183,7 +274,7 @@ public class AsymmetricView extends AbstractMatrixView {
             ArrayList<HBox> rowData = new ArrayList<>();
             for(int c=0; c<columns; c++) {
                 Pair<RenderMode, Object> item = template.get(r).get(c);
-                HBox cell = new HBox();  // wrap everything in an HBox so a border can be added easily
+                final HBox cell = new HBox();  // wrap everything in an HBox so a border can be added easily
                 Label label = null;
 
                 Background defaultBackground = DEFAULT_BACKGROUND;
@@ -198,8 +289,8 @@ public class AsymmetricView extends AbstractMatrixView {
                     case PLAIN_TEXT_V -> {
                         label = new Label((String) item.getValue());
                         label.setRotate(-90);
-                        cell.setAlignment(Pos.BOTTOM_RIGHT);
                         label.setPadding(new Insets(1));
+                        cell.setAlignment(Pos.BOTTOM_RIGHT);
                         Group g = new Group();  // label will be added to a group so that it will be formatted correctly if it is vertical
 
                         g.getChildren().add(label);
@@ -207,8 +298,8 @@ public class AsymmetricView extends AbstractMatrixView {
                     }
                     case ITEM_NAME -> {
                         label = new Label();
-                        label.textProperty().bind(((DSMItem) item.getValue()).getName());
                         label.setPadding(new Insets(0, 5, 0, 5));
+                        label.textProperty().bind(((DSMItem) item.getValue()).getName());
                         cell.setAlignment(Pos.CENTER_RIGHT);
                         label.setMinWidth(Region.USE_PREF_SIZE);
                         cell.getChildren().add(label);
@@ -217,7 +308,7 @@ public class AsymmetricView extends AbstractMatrixView {
                             if (e.getButton().equals(MouseButton.PRIMARY)) {
                                 editItemName(((DSMItem) item.getValue()).getUid());
                                 grid.resizeColumn(finalC);
-                                grid.updateGrid();
+                                grid.resizeRow(0);  // 0 is first row which contains the vertical names (this is cheating)
                             }
                         });
                     }
@@ -236,11 +327,11 @@ public class AsymmetricView extends AbstractMatrixView {
                             if (e.getButton().equals(MouseButton.PRIMARY)) {
                                 editItemName(((DSMItem) item.getValue()).getUid());
                                 grid.resizeRow(finalR);
-                                grid.updateGrid();
+                                grid.resizeColumn(1);  // 1 is first column which contains the horizontal names (this is cheating)
                             }
                         });
                         cell.setMinWidth(maxHeight);  // set a min width so that the matrix is less boxy (all connection items will follow this even if not
-                        // explicitly set due to how the freeze grid is set up)
+                                                      // explicitly set due to how the freeze grid is set up)
                     }
                     case GROUPING_ITEM -> {  // dropdown box for choosing group
                         ComboBox<Grouping> groupings = new ComboBox<>();
@@ -252,7 +343,7 @@ public class AsymmetricView extends AbstractMatrixView {
                         groupings.setCellFactory(groupingItemCellFactory);
                         groupings.setButtonCell(groupingItemCellFactory.call(null));
 
-                        groupings.getItems().addAll(matrix.getGroupings(true));
+                        groupings.getItems().addAll(matrix.getGroupings());
                         groupings.getSelectionModel().select(((DSMItem) item.getValue()).getGroup1());
                         groupings.getSelectionModel().selectedItemProperty().addListener((obs, oldVal, newVal) -> {
                             matrix.setItemGroup((DSMItem) item.getValue(), groupings.getValue());
@@ -264,31 +355,6 @@ public class AsymmetricView extends AbstractMatrixView {
 
                         cell.getChildren().add(groupings);
                     }
-                    case GROUPING_ITEM_V -> {  // dropdown box for choosing group
-                        ComboBox<Grouping> groupings = new ComboBox<>();
-                        groupings.setMinWidth(Region.USE_PREF_SIZE);
-                        groupings.setPadding(new Insets(0));
-                        groupings.setBackground(new Background(new BackgroundFill(Color.TRANSPARENT, new CornerRadii(3), new Insets(0))));
-                        groupings.setRotate(-90);
-
-                        Callback<ListView<Grouping>, ListCell<Grouping>> groupingItemCellFactory = getGroupingDropDownFactory(groupings);
-                        groupings.setCellFactory(groupingItemCellFactory);
-                        groupings.setButtonCell(groupingItemCellFactory.call(null));
-
-                        groupings.getItems().addAll(matrix.getGroupings(false));
-                        groupings.getSelectionModel().select(((DSMItem) item.getValue()).getGroup1());
-                        groupings.getSelectionModel().selectedItemProperty().addListener((obs, oldVal, newVal) -> {
-                            matrix.setItemGroup((DSMItem) item.getValue(), groupings.getValue());
-                            matrix.setCurrentStateAsCheckpoint();
-                            for (Cell c_ : cells) {
-                                refreshCellHighlight(c_);
-                            }
-                        });
-
-                        Group g = new Group();  // box will be added to a group so that it will be formatted correctly if it is vertical
-                        g.getChildren().add(groupings);
-                        cell.getChildren().add(g);
-                    }
                     case INDEX_ITEM -> {
                         NumericTextField entry = new NumericTextField(((DSMItem) item.getValue()).getSortIndex());
                         entry.setPrefColumnCount(3);  // set size to 3 characters fitting
@@ -297,9 +363,10 @@ public class AsymmetricView extends AbstractMatrixView {
                         cell.setAlignment(Pos.CENTER);
                         int finalR = r;
                         int finalC = c;
-                        entry.setOnAction(e -> {
+                        entry.setOnAction(e -> {  // remove focus when enter key is pressed
                             cell.getParent().requestFocus();
                         });
+
                         entry.focusedProperty().addListener((obs, oldVal, newVal) -> {
                             if (!newVal) {  // if changing to not focused
                                 if (entry.getNumericValue() != null) {
@@ -314,33 +381,46 @@ public class AsymmetricView extends AbstractMatrixView {
                         });
                         cell.getChildren().add(entry);
                     }
+                    case UNEDITABLE_CONNECTION -> defaultBackground = UNEDITABLE_CONNECTION_BACKGROUND;
                     case EDITABLE_CONNECTION -> {
                         int rowUid = ((Pair<DSMItem, DSMItem>) item.getValue()).getKey().getUid();
                         int colUid = ((Pair<DSMItem, DSMItem>) item.getValue()).getValue().getUid();
                         label = getEditableConnectionCell(cell, locationLabel, rowUid, colUid, r, c);
+                        int finalR = r;
+                        int finalC = c;
+                        cell.addEventHandler(CELL_CHANGED_EVENT, e -> {
+                            symmetryHighlightCell(getCellByLoc(new Pair<>(finalR, finalC)));
+                        });
                     }
+
                 }
-                cell.setBorder(new Border(new BorderStroke(Color.BLACK, BorderStrokeStyle.SOLID, CornerRadii.EMPTY, BorderWidths.DEFAULT)));
+
                 cell.setPadding(new Insets(0));
-                //cell.styleProperty().bind(Bindings.concat("-fx-font-size: ", fontSize.asString(), "};"));
                 rowData.add(cell);
 
                 Cell cellObject = new Cell(new Pair<>(r, c), cell, label, fontSize);
+                cellObject.setCellBorder(Color.BLACK);
                 cellObject.updateHighlightBG(defaultBackground, "default");
                 cells.add(cellObject);
-
             }
             gridData.add(rowData);
         }
-        for(Cell cell : cells) {
-            refreshCellHighlight(cell);
+
+        // only run this section if symmetryValidation is enabled because it cells default to not being highlighted
+        // and it is a waste of precious time to clear something that is already cleared
+        if(symmetryValidation) {
+            symmetryHighlightAllCells();
+        }
+
+        for (Cell c_ : cells) {  // this is needed outsize the render loop so that the groupings and item names are highlighted correctly
+            refreshCellHighlight(c_);
         }
 
         grid.setGridDataHBox(gridData);
         grid.setFreezeLeft(3);
         grid.setFreezeHeader(2);  // freeze top two rows for symmetric matrix
 
-        ArrayList<Integer> importantRows = new ArrayList<>(Arrays.asList(0, 1, 2, 3));
+        ArrayList<Integer> importantRows = new ArrayList<>(Arrays.asList(0, 1, 2));
         ArrayList<Integer> importantCols = new ArrayList<>(Arrays.asList(0, 1, 2, 3));
         grid.resizeGrid(true, importantRows, importantCols);
         grid.updateGrid();
@@ -350,7 +430,8 @@ public class AsymmetricView extends AbstractMatrixView {
 
 
     /**
-     * sets the current rootLayout to an immutable view
+     * Creates the guid that displays a matrix in a static read only view. Uses the SymmetricDSMData's getGridArray() method
+     * to create the grid.
      */
     @Override
     protected void refreshStaticView() {
@@ -419,21 +500,13 @@ public class AsymmetricView extends AbstractMatrixView {
                         label.setMinWidth(Region.USE_PREF_SIZE);
                         cell.getChildren().add(label);
                     }
-                    case GROUPING_ITEM_V -> {
-                        label = new Label(((DSMItem) item.getValue()).getGroup1().getName());
-                        label.setRotate(-90);
-                        cell.setAlignment(Pos.BOTTOM_RIGHT);
-                        label.setMinWidth(Region.USE_PREF_SIZE);
-                        Group g = new Group();  // label will be added to a group so that it will be formatted correctly if it is vertical
-                        g.getChildren().add(label);
-                        cell.getChildren().add(g);
-                    }
                     case INDEX_ITEM -> {
                         label = new Label(String.valueOf(((DSMItem) item.getValue()).getSortIndex()));
                         cell.setAlignment(Pos.BOTTOM_RIGHT);
                         label.setMinWidth(Region.USE_PREF_SIZE);
                         cell.getChildren().add(label);
                     }
+                    case UNEDITABLE_CONNECTION -> defaultBackground = UNEDITABLE_CONNECTION_BACKGROUND;
                     case EDITABLE_CONNECTION -> {
                         int rowUid = ((Pair<DSMItem, DSMItem>) item.getValue()).getKey().getUid();
                         int colUid = ((Pair<DSMItem, DSMItem>) item.getValue()).getValue().getUid();
@@ -516,33 +589,27 @@ public class AsymmetricView extends AbstractMatrixView {
         HBox testCell = new HBox();  // wrap everything in an HBox so a border can be added easily
         testCell.getChildren().add(testLabel);
 
-        // Start by calculating the width of column 1 and height of row 1 (groupings)
-        ArrayList<Grouping> allGroupings = new ArrayList<>(matrix.getGroupings(true));
-        allGroupings.addAll(matrix.getGroupings(false));
-        String longestGroupingName = allGroupings.stream().max(Comparator.comparing(g -> g.getName().length())).orElse(new Grouping("", Color.BLACK)).getName();
+        // Start by calculating the width of column 1 (groupings)
+        String longestGroupingName = matrix.getGroupings().stream().max(Comparator.comparing(g -> g.getName().length())).orElse(new Grouping("", Color.BLACK)).getName();
         longestGroupingName = longestGroupingName.length() > "Groupings".length() ? longestGroupingName: "Groupings";
         testLabel.setText(longestGroupingName);
         double col1Width = Misc.calculateNodeSize(testCell).getWidth();
-        double row1Height = col1Width;
 
-        // calculate width of column 2 and height of row 2 (item names)
-        ArrayList<DSMItem> items = new ArrayList<>();
-        items.addAll(matrix.getRows());
-        items.addAll(matrix.getCols());
-        items.add(new DSMItem(-1.0, "Column Items"));
-        String longestItemName = items.stream().max(Comparator.comparing((DSMItem item) -> item.getName().toString().length())).orElse(new DSMItem(-1.0, "")).getName().getValue();
+        // calculate width of column 2 and height of row 1 (item names)
+        String longestItemName = matrix.getRows().stream().max(Comparator.comparing((DSMItem item) -> item.getName().toString().length())).orElse(new DSMItem(-1.0, "")).getName().getValue();
+        longestItemName = longestItemName.length() > "Column Items".length() ? longestItemName: "Column Items";
         testLabel.setText(longestItemName);
         double col2Width = Misc.calculateNodeSize(testCell).getWidth();
-        double row2Height = col2Width;
+        double row1Height = col2Width;
 
         // calculate width of column 3
         testLabel.setText("Re-Sort Index");
         double col3Width = Misc.calculateNodeSize(testCell).getWidth();
 
-        // calculate height of row 3 and connection cell size
+        // calculate height of row 2 and connection cell size
         testLabel.setText("x");
-        double row3Height = Misc.calculateNodeSize(testCell).getHeight();
-        double cellSize = row3Height;
+        double row2Height = Misc.calculateNodeSize(testCell).getHeight();
+        double cellSize = row2Height;
 
 
         // with the way freeze grid handles spanning cells there needs to be a bunch of nulls and a single
@@ -557,6 +624,8 @@ public class AsymmetricView extends AbstractMatrixView {
         // Also keep track of the uneditable connection cells so they can be
         // plotted as a box
         ArrayList<Triplet<Integer, Integer, Color>> connectionLocations = new ArrayList<>();
+        ArrayList<Pair<Integer, Integer>> uneditableConnectionLocations = new ArrayList<>();
+
 
         // update the cells so they can be displayed by the freeze grid
         for(int r=0; r<numRows; r++) {
@@ -565,6 +634,8 @@ public class AsymmetricView extends AbstractMatrixView {
                 Pair<RenderMode, Object> item = template.get(r).get(c);
                 HBox cell = new HBox();  // wrap everything in an HBox so a border can be added easily
                 Label label = null;
+
+                Background defaultBackground = DEFAULT_BACKGROUND;
 
                 switch (item.getKey()) {
                     case PLAIN_TEXT -> {
@@ -617,22 +688,6 @@ public class AsymmetricView extends AbstractMatrixView {
                         cell.setAlignment(Pos.CENTER);
                         cell.getChildren().add(label);
                     }
-                    case GROUPING_ITEM_V -> {  // dropdown box for choosing group
-                        Grouping group = ((DSMItem) item.getValue()).getGroup1();
-                        label = new Label(group.getName());
-                        label.setRotate(-90);
-                        label.setPadding(new Insets(0, 5, 0, 5));
-                        label.setTextFill(group.getFontColor());
-                        label.setMinWidth(Region.USE_PREF_SIZE);
-                        cell.setAlignment(Pos.CENTER);
-                        cell.getChildren().add(label);
-
-                        // set a width so that the matrix is less boxy (all connection items will follow this even if not
-                        // explicitly set due to how the freeze grid is set up)
-                        cell.setMinWidth(cellSize);
-                        cell.setMaxWidth(cellSize);
-                        cell.setPrefWidth(cellSize);
-                    }
                     case INDEX_ITEM -> {
                         Grouping group = ((DSMItem) item.getValue()).getGroup1();
                         label = new Label(String.valueOf(((DSMItem) item.getValue()).getSortIndex()));
@@ -642,12 +697,15 @@ public class AsymmetricView extends AbstractMatrixView {
                         cell.setAlignment(Pos.CENTER);
                         cell.getChildren().add(label);
                     }
-                    case EDITABLE_CONNECTION -> {
+                    case UNEDITABLE_CONNECTION -> {
+                        uneditableConnectionLocations.add(new Pair<>(r, c));
+                        cell = null;
                         if(connectionsRow == -1) {  // if not set, this will be the first connection cell so set it here
                             connectionsRow = r;
                             connectionsCol = c;
                         }
-
+                    }
+                    case EDITABLE_CONNECTION -> {
                         int rowUid = ((Pair<DSMItem, DSMItem>) item.getValue()).getKey().getUid();
                         int colUid = ((Pair<DSMItem, DSMItem>) item.getValue()).getValue().getUid();
                         DSMConnection conn = matrix.getConnection(rowUid, colUid);
@@ -680,7 +738,7 @@ public class AsymmetricView extends AbstractMatrixView {
                     Cell cellObject = new Cell(new Pair<>(r, c), cell, label, scaledFontSize);
                     rowData.add(cell);
                     cellObject.setCellBorder(Color.BLACK);
-                    cellObject.updateHighlightBG(DEFAULT_BACKGROUND, "default");
+                    cellObject.updateHighlightBG(defaultBackground, "default");
                     cells.add(cellObject);
                 }
             }
@@ -707,24 +765,25 @@ public class AsymmetricView extends AbstractMatrixView {
         double y = 0;
         for(DSMItem rowItem : matrix.getRows()) {
             for(DSMItem colItem : matrix.getCols()) {
-                // upper left triangle will be row color
-                graphics_context.setFill(rowItem.getGroup1().getColor());
-                double[] leftXCoordinates = {x, x + cellSize, x};
-                double[] leftYCoordinates = {y, y, y + cellSize};
-                graphics_context.fillPolygon(leftXCoordinates, leftYCoordinates, 3);
-
-                // lower right triangle will be column color
-                graphics_context.setFill(colItem.getGroup1().getColor());
-                double[] rightXCoordinates = {x, x + cellSize, x + cellSize};
-                double[] rightYCoordinates = {y + cellSize, y + cellSize, y};
-                graphics_context.fillPolygon(rightXCoordinates, rightYCoordinates, 3);
-
+                if(rowItem.getGroup1().equals(colItem.getGroup1())) {
+                    graphics_context.setFill(rowItem.getGroup1().getColor());
+                } else {
+                    graphics_context.setFill(defaultColor);
+                }
+                graphics_context.fillRect(x, y, cellSize, cellSize);
                 x += cellSize;
             }
             x = 0;  // reset to first column at start of canvas
             y += cellSize;  // move to next row
         }
 
+        // draw the uneditable connection boxes
+        graphics_context.setFill(UNEDITABLE_CONNECTION_BACKGROUND.getFills().get(0).getFill());
+        for(Pair<Integer, Integer> coordinates : uneditableConnectionLocations) {
+            x = (coordinates.getKey() - connectionsRow) * cellSize;  // coordinates offset by connections start index so subtract it
+            y = (coordinates.getValue() - connectionsCol) * cellSize;
+            graphics_context.fillRect(x, y, cellSize, cellSize);
+        }
 
         // draw the connections
         for(Triplet<Integer, Integer, Color> connection : connectionLocations) {
@@ -740,7 +799,7 @@ public class AsymmetricView extends AbstractMatrixView {
         gridData.get(connectionsRow).set(connectionsCol, canvasPane);  // update the item to be the canvas
         //endregion
 
-        for (Cell c_ : cells) {  // this is needed outsize the render loop so that the groupings and item names are highlighted correctly
+        for (Cell c_ : cells) {  // this is needed outside the render loop so that the groupings and item names are highlighted correctly
             refreshCellHighlight(c_);
         }
 
@@ -751,12 +810,11 @@ public class AsymmetricView extends AbstractMatrixView {
 
         // set up the freezing
         grid.setFreezeLeft(3);
-        grid.setFreezeHeader(3);
+        grid.setFreezeHeader(2);
 
         // manually size all the items to save time and be more accurate
         grid.setRowPrefHeight(0, row1Height);
         grid.setRowPrefHeight(1, row2Height);
-        grid.setRowPrefHeight(2, row3Height);
         grid.setColPrefWidth(0, col1Width);
         grid.setColPrefWidth(1, col2Width);
         grid.setColPrefWidth(2, col3Width);
@@ -772,4 +830,5 @@ public class AsymmetricView extends AbstractMatrixView {
 
         rootLayout.getChildren().addAll(grid.getGrid());
     }
+
 }
