+++ src/main/java/UI/SideBarViews/AbstractSideBar.java
@@ -1,12 +1,10 @@
-package Matrices.SideBarTools;
+package UI.SideBarViews;
 
 import Matrices.Data.AbstractDSMData;
 import Matrices.Data.Entities.DSMConnection;
 import Matrices.Data.Entities.DSMInterfaceType;
 import Matrices.Data.Entities.DSMItem;
-import Matrices.Data.Entities.Grouping;
-import Matrices.Data.SymmetricDSMData;
-import Matrices.Views.AbstractMatrixView;
+import UI.MatrixViews.AbstractMatrixView;
 import UI.Widgets.Misc;
 import javafx.beans.property.SimpleStringProperty;
 import javafx.beans.property.StringProperty;
@@ -20,8 +18,7 @@ import javafx.stage.Modality;
 import javafx.stage.Stage;
 import javafx.util.Callback;
 
-import java.util.Map;
-import java.util.Vector;
+import java.util.List;
 
 
 /**
@@ -171,7 +168,8 @@ public abstract class AbstractSideBar {
         addMatrixItems.setOnAction(e -> addMatrixItemsCallback());
         addMatrixItems.setMaxWidth(Double.MAX_VALUE);
 
-        deleteMatrixItems.setOnAction(e -> deleteMatrixItemsCallback());
+        // default to rows only for deleting items
+        deleteMatrixItems.setOnAction(e -> deleteMatrixItemsCallback(matrix.getRows()));
         deleteMatrixItems.setMaxWidth(Double.MAX_VALUE);
 
         appendConnections.setOnAction(e -> appendConnectionsCallback());
@@ -215,7 +213,94 @@ public abstract class AbstractSideBar {
     /**
      * Sets up the button callback for deleting items from the matrix
      */
-    protected abstract void deleteMatrixItemsCallback();
+    protected void deleteMatrixItemsCallback(List<DSMItem> items) {
+        Stage window = new Stage();  // Create Root window
+        window.initModality(Modality.APPLICATION_MODAL); //Block events to other windows
+        window.setTitle("Delete Rows/Columns");
+
+        // Create changes view and button for it
+        Label label = new Label("Changes to be made");
+        ListView<Integer> changesToMakeView = new ListView<>();
+        changesToMakeView.getSelectionModel().setSelectionMode(SelectionMode.MULTIPLE);
+        changesToMakeView.setCellFactory(param -> new ListCell<>() {  // item uid
+            @Override
+            protected void updateItem(Integer item, boolean empty) {
+                super.updateItem(item, empty);
+
+                if (empty || item == null) {
+                    setText(null);
+                } else {
+                    setText(matrix.getItem(item).getName().getValue());
+                }
+            }
+        });
+
+        Button deleteSelected = new Button("Delete Selected from Change Stack");
+        deleteSelected.setOnAction(e -> {
+            changesToMakeView.getItems().removeAll(changesToMakeView.getSelectionModel().getSelectedItems());
+        });
+
+        // Create user input area
+        HBox entryArea = new HBox();
+
+        // ComboBox to choose which row or column to modify connections of
+        ComboBox<DSMItem> itemSelector = new ComboBox<>();  // rowUid | colUid | name | weight
+        itemSelector.setCellFactory(param -> new ListCell<>() {
+            @Override
+            protected void updateItem(DSMItem item, boolean empty) {
+                super.updateItem(item, empty);
+
+                if (empty || item == null) {
+                    setText(null);
+                } else {
+                    setText(item.getName().getValue());
+                }
+            }
+        });
+        itemSelector.setMaxWidth(Double.MAX_VALUE);
+        HBox.setHgrow(itemSelector, Priority.ALWAYS);
+        itemSelector.setPromptText("Item Name");
+        itemSelector.getItems().addAll(items);
+
+        Button deleteItem = new Button("Delete Item");
+        deleteItem.setOnAction(e -> {
+            changesToMakeView.getItems().add(itemSelector.getValue().getUid());
+        });
+
+        entryArea.getChildren().addAll(itemSelector, deleteItem);
+        entryArea.setPadding(new Insets(10, 10, 10, 10));
+        entryArea.setSpacing(20);
+
+        // create HBox for user to close with our without changes
+        HBox closeArea = new HBox();
+        Button applyButton = new Button("Apply Changes");
+        applyButton.setOnAction(e -> {
+            for(Integer uid : changesToMakeView.getItems()) {
+                DSMItem item = matrix.getItem(uid);  // null check in case user tries to delete the same item twice
+                matrix.deleteItem(item);
+            }
+            window.close();
+            matrixView.refreshView();
+            matrix.setCurrentStateAsCheckpoint();
+        });
+
+        Button cancelButton = new Button("Cancel");
+        cancelButton.setOnAction(e -> {
+            window.close();
+        });
+        closeArea.getChildren().addAll(cancelButton, Misc.getHorizontalSpacer(), applyButton);
+
+        VBox layout = new VBox(10);
+        layout.getChildren().addAll(label, changesToMakeView, deleteSelected, entryArea, Misc.getVerticalSpacer(), closeArea);
+        layout.setAlignment(Pos.CENTER);
+        layout.setPadding(new Insets(10, 10, 10, 10));
+        layout.setSpacing(10);
+
+        //Display window and wait for it to be closed before returning
+        Scene scene = new Scene(layout, 700, 350);
+        window.setScene(scene);
+        window.showAndWait();
+    }
 
 
     /**
@@ -234,6 +319,8 @@ public abstract class AbstractSideBar {
      * Sets up the button callback for deleting connections in the matrix
      */
     protected abstract void deleteConnectionsCallback();
+
+
 
 
     /**
